requirements.txt
altair==5.5.0
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.0
gitdb==4.0.11
gitpython==3.1.43
idna==3.10
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
markupsafe==3.0.2
matplotlib==3.9.3
mdurl==0.1.2
narwhals==1.15.1
numpy==2.1.3
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.0.0
plotly==5.24.1
protobuf==5.29.0
pyarrow==18.1.0
pydeck==0.9.1
pygments==2.18.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyyaml==6.0.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
smmap==5.0.1
statsmodels==0.14.4
streamlit==1.40.2
tenacity==9.0.0
threadpoolctl==3.5.0
toml==0.10.2
tornado==6.4.2
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
watchdog==6.0.0
xlsxwriter==3.2.0
